requirements.txt
amqp==5.3.1
anyio==3.7.1
asks==3.0.0
async-generator==1.10
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
docx2txt==0.8
et_xmlfile==2.0.0
exceptiongroup==1.2.2
Flask==3.1.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
mega.py==1.0.8
numpy==2.0.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pluggy==1.5.0
prompt_toolkit==3.0.48
pycryptodome==3.21.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tenacity==5.1.5
tomli==2.2.1
trio==0.28.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
